split RaftStoreProxyFFIHelper and RaftStoreProxy, gather all ffi dep
