+++ engine_store_ffi/src/ffi/read_index_helper.rs
