+++ new-mock-engine-store/src/mock_cluster.rs
@@ -16,7 +16,7 @@ pub use engine_store_ffi::ffi::{
     interfaces_ffi::{
         EngineStoreServerHelper, RaftProxyStatus, RaftStoreProxyFFIHelper, RawCppPtr,
     },
-    UnwrapExternCFunc,
+    RaftStoreProxy, UnwrapExternCFunc,
 };
 pub use engine_store_ffi::TiFlashEngine;
 use engine_tiflash::DB;
@@ -180,7 +180,7 @@ impl<T: Simulator<TiFlashEngine>> Cluster<T> {
             key_manager: key_mgr.clone(),
             read_index_client: match router {
                 Some(r) => Some(Box::new(
-                    engine_store_ffi::read_index_helper::ReadIndexClient::new(
+                    engine_store_ffi::ffi::read_index_helper::ReadIndexClient::new(
                         r.clone(),
                         SysQuota::cpu_cores_quota() as usize * 2,
                     ),
@@ -190,7 +190,8 @@ impl<T: Simulator<TiFlashEngine>> Cluster<T> {
             kv_engine: std::sync::RwLock::new(Some(engines.kv.clone())),
         });
 
-        let mut proxy_helper = Box::new(RaftStoreProxyFFIHelper::new(&proxy));
+        let proxy_ref = proxy.as_ref();
+        let mut proxy_helper = Box::new(RaftStoreProxyFFIHelper::new(proxy_ref.into()));
         let mut engine_store_server = Box::new(EngineStoreServer::new(id, Some(engines)));
         engine_store_server.proxy_compat = proxy_compat;
         engine_store_server.mock_cfg = mock_cfg;
@@ -418,7 +419,7 @@ impl<T: Simulator<TiFlashEngine>> Cluster<T> {
             let mut lock = self.ffi_helper_set.lock().unwrap();
             let ffi_helper_set = lock.get_mut(&node_id).unwrap();
             ffi_helper_set.proxy.read_index_client = Some(Box::new(
-                engine_store_ffi::read_index_helper::ReadIndexClient::new(
+                engine_store_ffi::ffi::read_index_helper::ReadIndexClient::new(
                     router.clone(),
                     SysQuota::cpu_cores_quota() as usize * 2,
                 ),
